requirements.txt
stockfish~=3.28.0
Flask~=3.0.3
chess~=1.11.1
python-dotenv~=1.0.1